bug-fix: publication date
The publication dates were unified to comply with the date given on the publication webpage, typically as: "published: DATE" or "published online: DATE".

Previously, the dates relied on the automatic Zotero import of the .ris file provided by the webpage. However, this import was found unreliable (probably could be declared a bug): For some publishers, Zotero imports a value quite different to the date given by the webpage html. For example, for Nature Communications, Zotero finds correctly only the year and supplements it with a fake month, most often (always?) December, instead of the correct month and day, which is ignored. It looks as if the .ris field "PY" is used instead of "DA" or something similar (issue-date instead of publication-date?). It is hard to say which party is guilty for this behavior.

Many of the resulting differences were within few weeks, which is unimportant for the figures (changes would be hard to spot), but the difference can not be ignored for publications in Nature Publishing (Nat. Commun. npj Quant. Inf., Commun. Phys., ...), where it can become as large as almost a year (the true date is in January, replaced by a fake date in December).