database: rounding rehaul
Previously, rounding to 2 significant digits was done on export to TeX.

Now, there are two types of rounding: on import (input) and on display (output). The rounding on output is a task left for the TeX package siunitx. Its style is specified in the table header and can be changed without the need to change the values in the database themselves.

The rounding on input is as follows: when numbers are copied from articles, such as "100" or "0.1", they are interpreted as real values with precision implicit in the number of significant digits provided in the corresponding string. The above two strings would imply precision being 3 and 1 significant digits, respectively. This precision is kept when transforming the number. For example, the input "5 MHz" should not produce a fake many-digit precision when converted to time by formulas such as 1/2\pi f.

Finally, sinutix can handle also values with explicit precision, and combine that consistently with rounding. Even though precisions are not copied from articles at the moment, if they are in the future, it will be better to exploit siunitx, rather than developing the rules anew.